Add initial-value calculations, add tests
This is a fairly small addition to the simulation: before the first
step function, we run through all the accumulators and check to see if
they have any valid inbound `initial-value` links. If they do, we
average the initial values of all the source nodes of those links, and
set the accumulator's initial value to that. We then run the simulation
as normal.

By explicitly setting the initial value property, not only does the rest
of the simulation work just as normal, but the node's visible slider and
property dialog reflect the value as well (while in simulation mode).

Add tests that check: simple setting of initial value; averaging; and
the accumulator working as normal after initial value is set.