Tweak transfer node style slightly
If we have a transfer node, enlarge the node background image slightly
so it pushes up to the borders, make the inbound connection go slightly
under the node so it links without any white space, and tweak the image
to cut off the edges.

The net result is the connections seem to be pipes going into the pipe
transfer image.