[#155049999] Set CODAP property: preventBringToFront
CODAP will soon be adjusting the default setting for preventBringToFront
from true to false. This goes along with other changes intended to make
plugins behave more like other components in the CODAP workspace. Sage
relies on being placed in the background, especially in standalone mode,
so we initialize with the property, instead of relying on the default.