Ensure relationship always uses latest customData
Previously, if a user created a vary relationship, created some data, and
then drew a new graph for that relationship, the underlying relationship
would not update. This was because @updateCustomData was not being called
again if there was already data.

This could have been fixed by simply calling @updateCustomData for each
data point, and this is reasonably performant, but it seemed a wasteful
operation when the data transformation was so simple, and @dataPoints wasn't
being used elsewhere.

So now we just assume that @customData is always ordered, since it is, and
use it directly.

Update tests.

[#156227028]