Interpolate points between two quickly-drawn points
When the mouse moves quickly across the drawing canvas, we only see a small
subset of the events. With a simple technique of recording the x, y location
of the events and plotting them on the graph, we are left with lots of missing
points.

A previous method for addressing this seemed to be to set all the points
from x-4 to x+4 to the same value. However, (1) this caused a weird "plateau"
effect of a flat area of the graph, and (2) we still got holes most of the
time when dragging across the canvas, resulting in a tooth-shaped line.

Now we check the time between the last-drawn point and this one, and if it
is <100ms, we interpolate all the points linearly between them. This seems
to work much better.

Also the way we set data is simplified, from knowing that our data is ordered.

[#156226989]