Dragging beyond graph's x-bounds should still draw
When the user drags above and below the y-bounds of the graph, we already
treat it as if they are dragging at the very top or bottom, to make it
easier to draw maxes and mins.

This does the same thing with the x-bounds. When a user drags across the
canvas, the last x coordinate at the axis is likely to get dropped, making
it so that the line does not end cleanly at the graphs bounds. Now we capture
x values beyond the bounds, so we can cleanly set the ends of the line.

[#156226989]