Make Graph button a draggable attribute for CODAP
This sets the correct MIMEtype on the drag event so that CODAP can read
the event and accept an attribute drag.

The image is set to `draggable: true` to give us the default browser
handling of drag. Eventually we may want to use a custom image with the
attribute name.

[#144034943]