Use time-used experiments if we set time-based
Previously we only used time-based experiments if there was a collector
in the graph. This is now changed to look at the AppStore settings.

As part of this, it makes sense for the minigraphs to reset as the user
switches between modes, and standardize showing/clearing when the run panel
is opened and closed.

[#153599165]