Move sliders to the right
The "slider" class that was previously only added when the slider was
editable didn't seem to be doing anything anymore -- possibly it was an
old feature. Re-use the "slider" class to be able to fix slider position.
This also fixes a small bug where the position of the slider is not always
constant.

[#156374559]