Fix: final failing test …
After getting tests running for the first time in 4 months …
There was one test that was actually failing.

I am not sure what the `time` parameter in the test is about.
There seems to be an offset of `10` in the `time` field …

PT Story:

Mocha tests stopped running 4 months ago

Turns out that we haven't been running our automated tests for about 4 months. We left in a describe.only() in one of our mocha tests, which prevents other tests from running.

[#159348401]
https://www.pivotaltracker.com/story/show/159348401