Fix: Sage/CODAP  increment exp. # -- Add comments.
At the time of document creation in CODAP we don't always know the final language the document is going to be rendered in. For example, An author sets up a CODAP document with Sage, and some other CODAP plugins. Next, they make copies of this document for several languages. Regardless of the author's language setting, the experiment number column has the un-localized label. Later when an i18 user collects experiment data for the first time, we then rename the column from the default to that user's language. We could partially avoid this if data CODAP table attributes supported `titles` for localized names.

Story:
Experiment increment doesn't work in translated versions of SageModeler/CODAP

[#159214059]
https://www.pivotaltracker.com/story/show/159214059