Fixed clearing node/link selection [#168140645]
Removed selection notification in SelectionManager._clearSelection and insured correct _clearSelection/clearSelection calls are made so that notification is performed only once.

Previously when a node was selected when another was selected the selection was cleared and notifications were sent before the new node was selected causing the inspector panel to close itself.