Fix selected tab unsetting when image is selected
In the image browser view, previously if we selected an image, our
selected tab would always switch back to 0.

This is fixed by removing the optionally-stateful tab-index handling in 
TabbedPanelView, which was messy and had a couple unnoticed bugs, and
instead requiring the use of props.

(Note, this could have also been fixed simply by fixing the implementation
of `componentWillReceiveProps`, which fails to account for `selectedTabIndex`
being undefined, but I'm going to want to use the `onTabSelected` listener
in the next commit, at which point no wrapper will be using the stateful
version of TabbedPanelView, so it seems pointless to support it.)