Add imageCollections url param, load to palleteStore
If the `imageCollections` parameter is added to the url, we will load
one or more collections of images into the palleteStore. The user-facing
feature will be in the next commit.

Included are some demos for connected-bio (eventually will be populated
with real images). Note that although the images are added to the repo,
they are in assets/ and so do not increase the app size.

NB: Other projects or authors could host images and json files anywhere.