Don't call finish on image cancel
This fixes a pre-existing bug that was made worse with new collections
work.

Previously, if the user drags the New Image icon into the canvas, we
show the New Image Dialog, and when an image is selected, we auto-close
the dialog and invoke the call back to place the image.

Because of the way we invoke the callback and close the dialog in
`finish`, which was called by `cancel`, if the user viewed an image and
then pressed cancel to select another, the dialog would close.

This was made more obvious when the feature was added to cancel the
selected image when tabs were changed.

Update test to reflect use-case better.