Add select of app when embedded in CODAP [#173785724]
Sends CODAP a select message to select the app when any click is performed in the app.  This will unselect all other CODAP components which will also remove their window menus.  This is needed after the new text component added a large menu.

The click handler was added both to the main app view, attached to the window, and the graph view as the graph view uses jsPlumb which attaches its own click handler which prevents the outer window handler from being called.