Guard against LARA passing empty string
If a model has no authoredState, and the model is being used in runtime,
LARA will pass in an initial state as 

```
{
    ...
    authoredState: null
}
```

which is fine.

However, for some reason in report-model, LARA passes the initial state as

```
{
    ...
    authoredState: ""
}
```

which causes `JSON.parse` to throw an error. This adds a simple check on
length which should fix that, and also protects it further with try/catch.