Collapsed slider dragging actions into one undoable step [#171509259]
This adds both the wrapping of a command batch around the slider start/end drag and a new setter method in the undoRedo manager to enable/disable further command batches.  This needed as each slider value change caused a seperate command batch to be created.

This also fixes a typo in the internal end command batch method.